// intranet/src/app/core/helios-api-client/model/membreDTO.ts
/**
 * Helios API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CentreDTO } from './centreDTO';
import { TypeMembreDTO } from './typeMembreDTO';
import { CiviliteDTO } from './civiliteDTO';
import { StatutMembreDTO } from './statutMembreDTO';


export interface MembreDTO { 
    id: number;
    nom: string;
    prenom: string;
    dateNaissance: string | null;
    civilite: CiviliteDTO;
    typeMembre: TypeMembreDTO;
    centre: CentreDTO;
    statut: StatutMembreDTO;
    email: string | null;
    emailValide: boolean;
    telephone: string | null;
    portable: string | null;
    adresse: string | null;
    codePostal: string | null;
    ville: string | null;
    pays: string | null;
    commentaires: string | null;
    connaissances: string | null;
    profession: string | null;
}